fix some flow errors
